+++ src/js/components/presentationals/HomePresentational.js
@@ -58,13 +58,13 @@ export const HomePresentational = props => {
   );
 };
 
-const mapStateToProps = state => {
+export const mapStateToProps = (state) => {
   return {
     isLoading: state.login.isLoading
   };
 }
 
-const mapDispatchToProps = dispatch => {
+export const mapDispatchToProps = (dispatch) => {
   return {
     login: (data) => dispatch(loginAction.post(data))
   }