+++ src/js/reducers/app/login.js
@@ -3,7 +3,6 @@ const initialState = {
 };
 
 const login = (state = initialState, action) => {
-    console.log(action)
     switch (action.type) {
         case "TRY_POST_LOGIN": {
             return Object.assign({}, state, {